+++ FinalProjectAtm.Razor/FinalProjectAtm.Razor/Pages/Transactions/Withdrawal.cshtml.cs
@@ -35,7 +35,7 @@ namespace FinalProjectAtm.Razor.Pages.Transactions
             }
         }
 
-        public IActionResult OnPost(int amount)
+        public IActionResult OnPost(double amount)
         {
 
             if (amount <= 0)
